Interceptor

+++ src/main/java/com/example/demo/ib/IBApplication.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.example.demo.exception.ProductNotFoundException;
 import com.example.demo.model.Product;
 
 @RestController
@@ -33,12 +34,18 @@ public class IBApplication {
 
     @DeleteMapping("/products/{id}")
     public ResponseEntity<Object> deleteProduct(@PathVariable String id) {
+        if (!productRepo.containsKey(id)) {
+            throw new ProductNotFoundException();
+        }
         productRepo.remove(id);
         return new ResponseEntity<>("Product deleted", HttpStatus.OK);
     }
 
     @PutMapping("/products/{id}")
     public ResponseEntity<Object> updateProduct(@PathVariable String id, @RequestBody Product product) {
+        if (!productRepo.containsKey(id)) {
+            throw new ProductNotFoundException();
+        }
         productRepo.remove(id);
         product.setId(id);
         productRepo.put(id, product);
@@ -54,6 +61,12 @@ public class IBApplication {
     @GetMapping("/products")
     public ResponseEntity<Object> getProducts() {
         return new ResponseEntity<>(productRepo.values(), HttpStatus.OK);
+    }
+
+    @GetMapping("/products/{id}")
+    public ResponseEntity<Object> getProductsbyID(@PathVariable String id) {
+        Product result = productRepo.get(id);
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
 }
